languages and site separated to own route and moved to plugin
