+++ kirby/site/config/config.php
@@ -13,20 +13,5 @@ return [
         'allowInsecure' => true,
         'basicAuth' => true
     ],
-    'kirby-vue-starterkit.plugin.useVueIndex' => false,
-    'routes' => [
-        [
-            'pattern' => 'languages.json',
-            'action'  => function () {
-                return Response::json(array_values(kirby()->languages()->map(function ($language) {
-                    return [
-                        'code' => $language->code(),
-                        'title' => $language->name(),
-                        'prefix' => $language->path(),
-                        'isDefault' => $language->isDefault()
-                    ];
-                })->data()));
-            }
-        ]
-    ]
+    'kirby-vue-starterkit.plugin.useVueIndex' => false
 ];
