+++ kirby/site/plugins/kirby-vue-starterkit/index.php
@@ -10,5 +10,5 @@ Kirby::plugin('kirby-vue-starterkit/plugin', [
     ],
     'api' => array_merge(option('api', []), require_once __DIR__ . '/api.php'),
     'tags' => array_merge(option('tags', []), require_once __DIR__ . '/tags.php'),
-    'routes' => array_merge(option('routes', []), option('kirby-vue-starterkit.plugin.useVueIndex') ? require_once __DIR__ . '/routes.php' : [])
+    'routes' => array_merge(option('routes', []), require_once __DIR__ . '/routes.php')
 ]);
