+++ vue-json/src/composables/use-kirby.js
@@ -1,41 +1,27 @@
-import { modifyPageHtml } from '../utils/modify-page-html'
 import { useLanguage } from '../composables/use-language'
+import { HtmlUtils } from '../utils/html.utils'
+import { PathUtils } from '../utils/path.utils'
 
 export const useKirby = () => {
-  const baseUrl =
+  const kirbyUrl =
     process.env.NODE_ENV === 'production' ? process.env.VUE_APP_KIRBY_URL : window.location.origin + process.env.VUE_APP_BASE_URL.slice(0, -1)
 
-  const toPageId = path => {
-    if (path.startsWith('/')) path = path.slice(1)
-    if (path.endsWith('/')) path = path.slice(0, -1)
-
-    return path || 'home'
-  }
-
-  const getLanguages = async () => {
-    const resp = await fetch(`${baseUrl}/languages.json`)
-    const languages = await resp.json()
-
-    return languages
-  }
-
-  const getSite = async () => {
+  const getJson = async uri => {
     const { prefix } = useLanguage()
+    const resp = await fetch(`${kirbyUrl}/${prefix}/${uri}.json`)
 
-    const resp = await fetch(`${baseUrl}/${prefix}/home.json`)
-    const home = await resp.json()
-
-    return home.site
+    return await resp.json()
   }
+
+  const getLanguages = async () => await getJson('languages')
+
+  const getSite = async () => await getJson('site')
 
   const getPage = async path => {
-    const { prefix } = useLanguage()
-
-    const resp = await fetch(`${baseUrl}/${prefix}/${toPageId(path)}.json`)
-    const page = await resp.json()
+    const page = await getJson(PathUtils.toPageUri(path))
 
     // fix relative links
-    modifyPageHtml(page, html => {
+    HtmlUtils.modifyPageHtml(page, html => {
       for (const a of html.getElementsByTagName('a')) {
         a.href = a.href.replace(process.env.VUE_APP_KIRBY_URL, process.env.VUE_APP_BASE_URL.slice(0, -1))
       }
