+++ vue-json/src/utils/path.utils.js
@@ -1,0 +1,8 @@
+export const PathUtils = {
+  toPageUri: path => {
+    if (path.startsWith('/')) path = path.slice(1)
+    if (path.endsWith('/')) path = path.slice(0, -1)
+
+    return path || 'home'
+  }
+}
